+++ doc/release-notes.md
@@ -3,3 +3,5 @@ Bitcoin ABC version 0.18.6 is now available from:
   <https://download.bitcoinabc.org/0.18.6/>
 
 This release includes the following features and fixes:
+ - Add `getfinalizedblock` rpc to allow node operators to introspec
+ the current finalized block.